HYD-7283 GUI: Integrate prettier, update eslint rules, apply rule to all pages
We need to integrate prettier, eslint-config-prettier, and
eslint-plugin-prettier into GUI. Once integrated, eslint rules need to
be updated such that eslint agrees with prettier rules. Finally, we need
to apply the new rules to all files. This patch will integrate prettier
and apply the rule to all files.

+++ source/iml/login/login-controller.js
@@ -6,12 +6,19 @@
 import * as fp from 'intel-fp';
 
 import eulaTemplate from './assets/html/eula.html!text';
-import accessDeniedTemplate from '../access-denied/assets/html/access-denied.html!text';
+import accessDeniedTemplate
+  from '../access-denied/assets/html/access-denied.html!text';
 
-export default function LoginCtrl ($uibModal, $q, SessionModel, help, navigate, ALLOW_ANONYMOUS_READ) {
+export default function LoginCtrl(
+  $uibModal,
+  $q,
+  SessionModel,
+  help,
+  navigate,
+  ALLOW_ANONYMOUS_READ
+) {
   'ngInject';
-
-  function initializeEulaDialog (user) {
+  function initializeEulaDialog(user) {
     return $uibModal.open({
       template: eulaTemplate,
       controller: 'EulaCtrl',
@@ -24,16 +31,14 @@ export default function LoginCtrl ($uibModal, $q, SessionModel, help, navigate, 
     }).result;
   }
 
-  const initializeDeniedDialog = function initializeDeniedLoginFn () {
+  const initializeDeniedDialog = function initializeDeniedLoginFn() {
     return $uibModal.open({
       template: accessDeniedTemplate,
       controller: 'AccessDeniedCtrl',
       backdrop: 'static',
       keyboard: false,
       resolve: {
-        message: fp.always(
-          help.get('access_denied_eula')
-        )
+        message: fp.always(help.get('access_denied_eula'))
       }
     }).result;
   }.bind(this);
@@ -41,23 +46,27 @@ export default function LoginCtrl ($uibModal, $q, SessionModel, help, navigate, 
   /**
    * Submits the login, calling nextStep if successful.
    */
-  this.submitLogin = function submitLogin () {
+  this.submitLogin = function submitLogin() {
     this.inProgress = true;
 
-    this.validate = SessionModel.login(this.username, this.password).$promise
-    .then(function (session) {
-      return session.user.actOnEulaState(initializeEulaDialog, initializeDeniedDialog);
-    })
-    .then(() => navigate())
-    .catch(function (reason) {
-      if (reason === 'dismiss')
-        return SessionModel.delete().$promise;
+    this.validate = SessionModel.login(this.username, this.password)
+      .$promise.then(function(session) {
+        return session.user.actOnEulaState(
+          initializeEulaDialog,
+          initializeDeniedDialog
+        );
+      })
+      .then(() => navigate())
+      .catch(function(reason) {
+        if (reason === 'dismiss') return SessionModel.delete().$promise;
 
-      return $q.reject(reason);
-    })
-    .finally(function () {
-      this.inProgress = false;
-    }.bind(this));
+        return $q.reject(reason);
+      })
+      .finally(
+        function() {
+          this.inProgress = false;
+        }.bind(this)
+      );
   };
 
   this.ALLOW_ANONYMOUS_READ = ALLOW_ANONYMOUS_READ;
